Not perfectly working solution of show All courses for Teachr of type USER
It is an experimentation in progress. Currently it is wrong that main nav's courses changes to the specific's nav's courses. This will be fixed. It was really annoying with all the connections and forwards. But I build up some basic idea how it is supposed to be done and it will look better with time.