Now we should be able to create our table dynamically
With tableView fragment we should be able to send 2 sets of data - one list of table header and one list of database data. The table headers should always be there but the table data might be empty. In case it is empty it should show a hardcoded message.

OBS it is only tested with courses. and it could be seen in teacher from type user.