Fixes on the display of student courses
The display shows only courses that have a teacher! if teacher for a course is null the student will not see that course.
Small fixes in tableView to skip the first row that has the data about the result table headers. New method is created in MyPageServlet that will be preparing the ArrayList with strings that are later sent via Context to the JSP fragmet table view.