Dealing with submenu items in the GET
Now we will be working mostly in MyPage so we need to know from which submenu is the user coming and show the expcted info based on it. It is only done in teacher of type user. it works with the change in query show Students in rqueries (the first one) to show all students.