Right pathing and preparation
Now the whole structure for the rest of the code without main page Register is implemented. Error message is shown if a logged in user tries to reach a submeny item that does not exist. The 404 will be needed if they try to reach item that does not exist later but that is enough for now to ensure that student's wont try to reach teacher's submenues without being given information that this page does not exist (for them :))