New feature - report on course info
Now the user teacher and the admin teacher can see a report based on course name or course id. The report shows all teacher leading the course, and all students, also which is the course and how many points does it give. it could be found under menu item Course information.