Basic session handle
Now when our user attempt to add a course and the session has expired they will not succeed. Not elegant solution and it requires an action to actually reach it. Next step will be to just inform the user that their session has expired. But this might wait another course